+++ changes.md
@@ -1,13 +1,19 @@
-#Changes
+# Changes
 [TOC]
 Update 13.1.0 Bug fixes and some new features (Will upgrade version to 14.0.0 if significant changes are made)
 -------------
 Added: 
 - Connections:Lock() -- Prevents a connection object form being fired
 - Connections:Unlock() -- Removes the restriction imposed by conn:Lock()
+- 
 
 Fixed:
 - Minor bug with multi:newThread() in how names and functions were managed
+- Major bug with the system thread handler. Saw healthy threads as dead ones 
+- 
+
+Changed: 
+- getTasksDetails("t"), the table varaiant, formats threads, and system threads in the same way that tasks are formatted
 - 
 
 Update 13.0.0 Added some documentation, and some new features too check it out!